Change file including and use forward declarations
